Move initSelects() from utils.js to selects.js
